src/server/requirements.txt
marshmellow
sqlalchemy
flask
flask-jwt-extend
flask-jwt
